AddReferences method returns too early
The AddRefrences method exists in similar versions in several of the sample node manager classes. The method loops through a list of (external) references and adds them to the designated objects. When obtaining node handles it also checks that the node is valid in the current context. The problem was that it skipped the rest of  the nodes in the list as soon as it found one node that wass not valid. This change ensures that all nodes in the references list are processed.