Fixed UWP and Xamarin build (#574)
Fixed the failing Azure pipeline UWP build:
Changed generic EventHandler to now as PublishStateChangedEventHandler typed EventHandler. This fixed the errors in the UWP build.

Fixed the Xamarin build:
The msbuild for Android SDK 25 (Android 7.1) seems to be non-functional. Could not reproduce the exact same behavior locally - might also be caused by an error in the Azure pipeline configuration. (If the pipeline fails in the PR, I'll just remove the Xamarin related change...)